Create eigenObs.m
The method we built together in class. Renamed it as obsolete because it takes 8 minutes to run so I replaced it with a faster one, it's still the most accurate though so we might use it for the final thing (but like run the program an hour before we're due to start the presentation lol.